requirements.txt
aiohttp==3.9.3
email_validator==2.1.0.post1
fastapi==0.109.0
fastapi_utils==0.2.1
motor==3.3.2
pydantic==1.10.14
pymongo==4.6.2
python-dotenv==1.0.1
Requests==2.31.0
six==1.16.0
uvicorn==0.27.0.post1
itsdangerous==2.1.2
colored==2.2.4